move fetcher out of the engine.rs file
