+++ crates/indexer/src/lib.rs
@@ -7,3 +7,5 @@ mod test;
 pub mod engine;
 pub mod error;
 pub use engine::Engine;
+
+pub mod fetcher;
